make feed store specs protocol
