+++ FeedFeatureTests/Feed Cache/CodableFeedStoreTests.swift
@@ -8,37 +8,7 @@
 import XCTest
 import FeedFeature
 
-protocol FeedStoreSpecs {
-     func test_retrieve_deliversEmptyOnEmptyCache()
-     func test_retrieve_hasNoSideEffectsOnEmptyCache()
-     func test_retrieve_deliversFoundValuesOnNonEmptyCache()
-     func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
-
-     func test_insert_overridesPreviouslyInsertedCachedValues()
-
-     func test_delete_hasNoSideEffectsOnEmptyCache()
-     func test_delete_emptiesPreviouslyInsertedCache()
-
-     func test_storeSideEffects_runSerially()
-}
-
-protocol FailableRetrieveFeedStoreSpecs {
-    func test_retrieve_deliversFailureOnRetrievalError()
-    func test_retrieve_hasNoSideEffectOnFailure()
-}
-
-protocol FailableInsertFeedStoreSpecs {
-    func test_insert_deliversErrorOnInsertionError()
-    func test_insert_hasNoSideEffectsOnInsertionError()
-}
-
-protocol FailableDeletionFeedStoreSpecs {
-    func test_delete_deliversErrorOnDeletionError()
-    func test_delete_hasNoSideEffectOnDeletionError()
-}
-
-
-class CodableFeedStoreTests: XCTestCase {
+class CodableFeedStoreTests: XCTestCase, FailableFeedStoreSpecs {
 
     override func setUp() {
         super.setUp()
