+++ FeedFeatureTests/Feed Cache/FailableFeedStoreSpecs.swift
@@ -1,0 +1,39 @@
+//
+//  FailableFeedStoreSpecs.swift
+//  FeedFeatureTests
+//
+//  Created by Srinivasan Rajendran on 2020-12-27.
+//
+
+import Foundation
+
+protocol FeedStoreSpecs {
+     func test_retrieve_deliversEmptyOnEmptyCache()
+     func test_retrieve_hasNoSideEffectsOnEmptyCache()
+     func test_retrieve_deliversFoundValuesOnNonEmptyCache()
+     func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
+
+     func test_insert_overridesPreviouslyInsertedCachedValues()
+
+     func test_delete_hasNoSideEffectsOnEmptyCache()
+     func test_delete_emptiesPreviouslyInsertedCache()
+
+     func test_storeSideEffects_runSerially()
+}
+
+protocol FailableRetrieveFeedStoreSpecs: FeedStoreSpecs {
+    func test_retrieve_deliversFailureOnRetrievalError()
+    func test_retrieve_hasNoSideEffectOnFailure()
+}
+
+protocol FailableInsertFeedStoreSpecs: FeedStoreSpecs {
+    func test_insert_deliversErrorOnInsertionError()
+    func test_insert_hasNoSideEffectsOnInsertionError()
+}
+
+protocol FailableDeletionFeedStoreSpecs: FeedStoreSpecs {
+    func test_delete_deliversErrorOnDeletionError()
+    func test_delete_hasNoSideEffectOnDeletionError()
+}
+
+typealias FailableFeedStoreSpecs = FailableRetrieveFeedStoreSpecs & FailableInsertFeedStoreSpecs & FailableDeletionFeedStoreSpecs
